Extracting common envoy code into a package
